Refactor chart class structure
This makes the code more readable and easier to use and extend.
